ToStringBuilder redesigned a little.
The process of string building was divided into two phases.
At first, all of the information is tranformed to string and
added into a special buffer together with a control code.
During the second phase the spaces and new line characters are
added according to the control codes not to exceed line length.